Package proxy moved to the package external
